update repository package: add PlayerBotRepository, move repositories to appropriate subpackages
