+++ src/main/java/self/izouir/bitkionline/repository/player/PlayerRepository.java
@@ -1,9 +1,12 @@
-package self.izouir.bitkionline.repository;
+package self.izouir.bitkionline.repository.player;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import self.izouir.bitkionline.entity.Player;
+import self.izouir.bitkionline.entity.player.Player;
+
+import java.util.Optional;
 
 @Repository
 public interface PlayerRepository extends JpaRepository<Player, Long> {
+    Optional<Player> findByUsername(String username);
 }
